Booking tickets and managing reservations

+++ Common/Commands/RemoveReservation.cs
@@ -1,13 +1,12 @@
-﻿using System;
-using System.Collections.Generic;
 using System.Linq;
-using System.Text;
-using Administration.Messages;
+using System.Windows;
 using Caliburn.Micro;
 using Common.Infrastucture;
+using Common.Messages;
+using Common.Utils;
 using Connection;
 
-namespace Administration.Commands.Reservations
+namespace Common.Commands
 {
     public class RemoveReservation : ICommand
     {
@@ -27,9 +26,16 @@ namespace Administration.Commands.Reservations
 
         public void Execute()
         {
+            if (MessageBoxService.ShowConfirmationMessage() != MessageBoxResult.Yes)
+            {
+                return;
+            }
+
             DBConnection dbConnection = _connectionProvider.GetConnection();
+            
             RESERVATION reservationToRemove = dbConnection.RESERVATION
                 .Single(reservation => reservation.SYMBOL == _reservationToRemove.SYMBOL);
+            reservationToRemove.FLY.FREE_PLACES_NUMBER++;
             dbConnection.RESERVATION.DeleteObject(reservationToRemove);
             dbConnection.SaveChanges();
